Add files via upload
Now the fragment shader in ShaderNeural.js automatically update the number of network's parameters, retrieving the numbers from the "samples" and "planes" members of info.json. This is done using the function setShaderInfo, which is defined in ShaderNeural.js and called at initialization time in LayerNeuralRTI.js. The network is still implemented to use a fixed number of 3 layers.